CAT-319 Update UI validation to support custom organisation source
