+++ src/pages/index.ts
@@ -2,19 +2,16 @@ import Home from "./Home";
 import Profile from "./Profile";
 import ProfileUpdate from "./ProfileUpdate";
 import Users from "./Users";
-
-import {
-  RequestValidation,
-  Validations,
-  ValidationDetails,
-} from "./Validations";
+import RequestValidation from "./validations/RequestValidation";
+import ValidationDetails from "./validations/ValidationDetails";
+import ValidationList from "./validations/ValidationList";
 
 export {
   Home,
   Profile,
   Users,
   RequestValidation,
-  Validations,
+  ValidationList,
   ProfileUpdate,
   ValidationDetails,
 };
